# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Employee info api",
    "description": "employees",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/employees/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "Create a new Employee Info Id",
        "parameters":[
          {
            "name":"body",
            "in":"body",
            "required":true,
            "schema":{
              "type":"object",
              "properties":{
                "firstName":{
                  "type":"string"},
                  "lastName":{
                    "type":"string"},
                    "department":{
                      "type":"string"
                    },
                    "email":{
                      "type":"string"
                    },
                    "birthday":{
                      "type":"string"
                    }
                  }
                  }
                }
            
            
          
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/employees/{id}": {
      "put": {
        "description": "Update an existing contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
                "type": "string"
            },
            "description": "The ID of the contact to update"
        },
        {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "department": {
                        "type": "string"
                    },
                    "email":{
                      "type":"string"
                    },
                    "birthday":{
                      "type":"string"
                    }
                  }
                }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}